Delete condicional

+++ index.js
@@ -151,12 +151,15 @@ app.put('/usuarios/:id', (request, response) => {
 app.delete('/usuarios/:id', async (req, res) => {
     let conn;
         try{
+
+        var rows = await conn.query('SELECT * FROM INDIVIDUAL_USERS WHERE EMAIL = ?', [email, password]);
+            if(rows && rows.length)
             conn = await pool.getConnection();
             conn.query('DELETE FROM USERS WHERE email = ?', [req.params.email])
             .then((result) => {
-                conn.query('DELETE FROM COMPANIES WHERE email = ?', [req.params.email])
+                if(!(rows && rows.length)) conn.query('DELETE FROM COMPANIES WHERE email = ?', [req.params.email])
                     .then((result) => {
-                        conn.query('DELETE FROM INDIVIDUAL_USERS WHERE email = ?', [req.params.email])
+                        if(rows && rows.length) conn.query('DELETE FROM INDIVIDUAL_USERS WHERE email = ?', [req.params.email])
                             .then((result) => {
                                res.status(201).send('User deleted');
                             }).catch(err => {
